+++ src/math/webgl/binaryop_gpu.ts
@@ -15,8 +15,7 @@
  * =============================================================================
  */
 
-import * as util from '../../util';
-
+import * as broadcast_util from '../broadcast_util';
 import {GPGPUProgram} from './gpgpu_math';
 
 export const ADD = 'return a + b;';
@@ -38,7 +37,8 @@ export class BinaryOpProgram implements GPGPUProgram {
 
   constructor(op: string, aShape: number[], bShape: number[]) {
     this.params = [op];
-    this.outputShape = util.assertAndGetBroadcastShape(aShape, bShape);
+    this.outputShape =
+        broadcast_util.assertAndGetBroadcastShape(aShape, bShape);
     this.userCode = `
       float binaryOperation(float a, float b) {
         ${op}
